# doc/backend.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Backend API",
    "description": "The application backend api",
    "version": "1.0",
    "contact": {
      "name": "efly",
      "url": "https://efly-cetc.com",
      "email": "none@example.com"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "application/octet-stream"
  ],
  "produces": [
    "application/json",
    "application/octet-stream"
  ],
  "paths": {
    "/v1/applications": {
      "get": {
        "summary": "列表",
        "operationId": "AppService_Paginate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AppPaginateResponse"
            }
          },
          "201": {
            "description": "Returned when the resource created success.",
            "schema": {}
          },
          "204": {
            "description": "Returned when the resource deleted success.",
            "schema": {}
          },
          "400": {
            "description": "Returned when the request data validate failed.",
            "schema": {
              "$ref": "#/definitions/v1ErrorsObject"
            }
          },
          "401": {
            "description": "Returned when the user does not authentication.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource not found.",
            "schema": {}
          },
          "409": {
            "description": "Returned when request update conflict.",
            "schema": {}
          },
          "423": {
            "description": "Returned when resource locked.",
            "schema": {}
          },
          "429": {
            "description": "Returned when request too many in a unit time.",
            "schema": {}
          },
          "500": {
            "description": "Server error.",
            "schema": {}
          },
          "default": {
            "description": "Unexpect error.",
            "schema": {
              "$ref": "#/definitions/v1ErrorsObject"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "description": "Page number, default=1",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "limit",
            "description": "Page size, default=10",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "keyword",
            "description": "Keyword search, like:keyword=xx，keyword=id=1|name=xxx",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sort",
            "description": "Sort list, like: sort=+a,-b,+c,-d",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "AppService"
        ],
        "security": [
          {
            "AppId": [],
            "BearerToken": []
          }
        ],
        "x-order": 1
      },
      "post": {
        "summary": "创建",
        "operationId": "AppService_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1App"
            }
          },
          "201": {
            "description": "Returned when the resource created success.",
            "schema": {}
          },
          "204": {
            "description": "Returned when the resource deleted success.",
            "schema": {}
          },
          "400": {
            "description": "Returned when the request data validate failed.",
            "schema": {
              "$ref": "#/definitions/v1ErrorsObject"
            }
          },
          "401": {
            "description": "Returned when the user does not authentication.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource not found.",
            "schema": {}
          },
          "409": {
            "description": "Returned when request update conflict.",
            "schema": {}
          },
          "423": {
            "description": "Returned when resource locked.",
            "schema": {}
          },
          "429": {
            "description": "Returned when request too many in a unit time.",
            "schema": {}
          },
          "500": {
            "description": "Server error.",
            "schema": {}
          },
          "default": {
            "description": "Unexpect error.",
            "schema": {
              "$ref": "#/definitions/v1ErrorsObject"
            }
          }
        },
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AppFormData",
              "description": "data",
              "required": [
                "data"
              ]
            }
          }
        ],
        "tags": [
          "AppService"
        ],
        "security": [
          {
            "AppId": [],
            "BearerToken": []
          }
        ],
        "x-order": 3
      }
    },
...
    "/v1/config": {
      "get": {
        "summary": "配置信息",
        "description": "返回服务当前配置信息",
        "operationId": "ConfigService_Detail",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Config"
            }
          },
          "201": {
            "description": "Returned when the resource created success.",
            "schema": {}
          },
          "204": {
            "description": "Returned when the resource deleted success.",
            "schema": {}
          },
          "400": {
            "description": "Returned when the request data validate failed.",
            "schema": {
              "$ref": "#/definitions/v1ErrorsObject"
            }
          },
          "401": {
            "description": "Returned when the user does not authentication.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource not found.",
            "schema": {}
          },
          "409": {
            "description": "Returned when request update conflict.",
            "schema": {}
          },
          "423": {
            "description": "Returned when resource locked.",
            "schema": {}
          },
          "429": {
            "description": "Returned when request too many in a unit time.",
            "schema": {}
          },
          "500": {
            "description": "Server error.",
            "schema": {}
          },
          "default": {
            "description": "Unexpect error.",
            "schema": {
              "$ref": "#/definitions/v1ErrorsObject"
            }
          }
        },
        "tags": [
          "ConfigService"
        ],
        "security": [
          {
            "AppId": [],
            "BearerToken": []
          }
        ]
      },
      "put": {
        "summary": "更新配置",
        "description": "更新服务当前配置信息",
        "operationId": "ConfigService_Update",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Config"
            }
          },
          "201": {
            "description": "Returned when the resource created success.",
            "schema": {}
          },
          "204": {
            "description": "Returned when the resource deleted success.",
            "schema": {}
          },
          "400": {
            "description": "Returned when the request data validate failed.",
            "schema": {
              "$ref": "#/definitions/v1ErrorsObject"
            }
          },
          "401": {
            "description": "Returned when the user does not authentication.",
            "schema": {}
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {}
          },
          "404": {
            "description": "Returned when the resource not found.",
            "schema": {}
          },
          "409": {
            "description": "Returned when request update conflict.",
            "schema": {}
          },
          "423": {
            "description": "Returned when resource locked.",
            "schema": {}
          },
          "429": {
            "description": "Returned when request too many in a unit time.",
            "schema": {}
          },
          "500": {
            "description": "Server error.",
            "schema": {}
          },
          "default": {
            "description": "Unexpect error.",
            "schema": {
              "$ref": "#/definitions/v1ErrorsObject"
            }
          }
        },
        "parameters": [
          {
            "name": "debug",
            "description": "调试模式",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "ConfigService"
        ],
        "security": [
          {
            "AppId": [],
            "BearerToken": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "v1App": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "app id"
        },
        "name": {
          "type": "string",
          "description": "应用名称"
        },
        "type": {
          "$ref": "#/definitions/v1AppType",
          "description": "类型"
        },
        "description": {
          "type": "string",
          "description": "应用描述"
        },
        "scopes": {
          "type": "string",
          "description": "作用范围"
        },
        "package": {
          "type": "string",
          "description": "包名"
        },
        "project": {
          "type": "string",
          "description": "项目标识"
        },
        "manage": {
          "type": "boolean",
          "description": "是否管理型应用"
        },
        "enabled": {
          "$ref": "#/definitions/v1EnableState",
          "description": "启用状态"
        },
        "attrs": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "description": "应用属性"
        },
        "operateinfo": {
          "$ref": "#/definitions/v1OperateInfo",
          "description": "时间"
        },
        "conflict": {
          "$ref": "#/definitions/v1Conflict",
          "description": "冲突检测"
        }
      }
    },
    "v1AppFormData": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "应用名称"
        },
        "description": {
          "type": "string",
          "description": "应用描述"
        },
        "scopes": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "作用范围"
        },
        "type": {
          "$ref": "#/definitions/v1AppType",
          "description": "类型"
        },
        "package": {
          "type": "string",
          "description": "包名"
        },
        "project": {
          "type": "string",
          "description": "项目标识"
        },
        "manage": {
          "type": "boolean",
          "description": "是否管理型应用"
        },
        "attrs": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "description": "应用属性"
        }
      }
    },
    "v1AppPaginateDate": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "应用 id"
        },
        "name": {
          "type": "string",
          "description": "应用名称"
        },
        "type": {
          "$ref": "#/definitions/v1AppType",
          "description": "类型"
        },
        "manage": {
          "type": "boolean",
          "description": "是否管理型应用"
        },
        "enabled": {
          "$ref": "#/definitions/v1EnableState",
          "description": "启用状态"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        }
      }
    },
    "v1AppPaginateResponse": {
      "type": "object",
      "properties": {
        "meta": {
          "$ref": "#/definitions/v1PaginateMeta"
        },
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1AppPaginateDate"
          }
        }
      }
    },
    "v1AppType": {
      "type": "string",
      "enum": [
        "Backend",
        "Frontend"
      ],
      "default": "Backend"
    },
...
    "v1Config": {
      "type": "object",
      "properties": {
        "debug": {
          "type": "boolean",
          "description": "调试模式"
        }
      },
      "title": "TODO 8 change  to  project config and controller"
    },
...
    "v1Conflict": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "description": "The conflict value of the resource",
          "title": "Value"
        }
      },
      "title": "冲突检查值",
      "required": [
        "value"
      ]
    },
...
    "v1EnableState": {
      "type": "string",
      "enum": [
        "Disabled",
        "Enabled",
        "Default"
      ],
      "default": "Disabled",
      "title": "启用 禁用"
    },
    "v1ErrorObject": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "Response error code",
          "pattern": "^[0-9]{1,}$"
        },
        "message": {
          "type": "string",
          "description": "Response error message"
        }
      },
      "title": "错误"
    },
    "v1ErrorsObject": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "Response error code",
          "pattern": "^[0-9]{1,}$"
        },
        "message": {
          "type": "string",
          "description": "Response error message"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1ErrorObject"
          },
          "description": "Response errors",
          "title": "错误链"
        }
      },
      "title": "错误 多个错误"
    },
...
    "v1OperateInfo": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time",
          "description": "最后更新时间"
        },
        "deletedAt": {
          "type": "string",
          "format": "date-time",
          "description": "删除时间"
        }
      },
      "title": "操作信息"
    },
...
    "v1PaginateMeta": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer",
          "format": "int64",
          "description": "Resource total nums"
        },
        "page": {
          "type": "integer",
          "format": "int64",
          "description": "Current page id"
        },
        "limit": {
          "type": "integer",
          "format": "int64",
          "description": "Current page size"
        }
      },
      "title": "分页元数据"
    },
...
  },
  "securityDefinitions": {
    "AppId": {
      "type": "apiKey",
      "name": "X-App-Id",
      "in": "header"
    },
    "BearerToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}